requirements.txt
black==18.6b4
jikanpy==3.0.0
matplotlib==2.2.2
numpy==1.14.3
Pillow==5.1.0
rectangle-packer==1.1.0
